Fix sprite 0 hit check
Sprite 0 only counts when it is sprite 0 (out of 64) in the OAM, not
when it's the 0th sprite in the line. However, spr0 will always be the
0th sprite in the line, so if spr0's y value is in range, then it will
be at index 0 in OAM too.

This change also removes the old `spr0` variable: it's not needed, since
the sprite 0 flag cannot be cleared by any other mechanism, so there's
no need to track whether it has been set.