Fix aspect ratio, add overscan for top/bottom
The generated pixels are 256x240 (not including the border), but
typically only 256x224 pixels are visible (though it depends on the
display). In addition, the actual pixel aspect ratio is 8/7 =~ 1.143.
Since we expect our display to have square pixels, then we need to
stretch the image horizontally by this ratio (to 292.57x240).

We could clip the top and bottom rows afterward, but it's easier to
adjust the source UV to remove these rows, then adjust the desired
aspect ratio to take them into account (i.e. (256*8)/((240-16)*7) =~ 1.306).