Finally passes, but seems like a bit of a hack
Many changes to make it work more like the actual hardware

* Use len instead of len+1
* Use dmcen flag instead of dmcsilent
* The sample byte buffer can be in one of 3 states:
    0: no sample byte has been loaded
    1: a sample byte is loaded and is currently playing
    2: a sample byte is loaded but is not yet playing

Some interesting notes from analyzing Visual2A03:
* Nearly everything seems to be synchronized to the APU tick
* But writing to $4015 bit 4 enables PCM immediately
* Fetch only occurs if the sample buffer is empty. If you start the
  channel and a sample is already available, it will continue to play
  that sample before fetching the next one.
* Timer is 9 bit LFSR with taps at bits 4,8. It resets at $100. When it
  overflows, pcm_bit is incremented, pcm_sr is shifted, and frequency is
  reloaded (from $4010)
* When pcm_bit overflows from 7->0, then a memory fetch is triggered as
  long as the DMC channel is enabled
* When the fetch finishes the length counter is checked, if it is
  non-zero then it is decremented. Otherwise, the dmc_en flag will be
  cleared if the loop flag is not set. In either case the internal
  pcm_lc/pcm_a are reloaded.

apu_tests/7-dmc_basics #19 gave me a lot of trouble (as did others, it
seems, from reading forum.nesdev.com).

The test is a bit confusing: it starts a 17-byte sample (by loading 1
into $4013), then waits for the DMC enable flag to be cleared (line
174). Note that when the flag is cleared, the last fetched byte is still
being played. It then waits for 4 more "dmc_cycles" + 30 cpu ticks. By
this point the sample has long since finished. It then starts another
1-byte sample (i.e. loading 0 into $4013, see line 178). There is no
currently playing sample, and the sample buffer is empty, so a byte will
be fetched. However the fetched byte won't be loaded and start playing
until pcm_bits rolls over to 0. The test is written so this happens
immediately after the dmc_en flag is cleared. (The dmc_en flag is
cleared because pcm_lc is 0 after the fetch.)

At this point the dmc_en flag should be clear, but the fetched sample
will be playing. The test checks for this (see line 182). The test then
starts another 1-byte sample (line 183), which sets the dmc_en flag.
However, because the current sample is still playing, this sample byte
is not yet fetched. After the byte fetched at line 178 is finished, this
next byte is fetched, dmc_en is cleared, and the final byte continues to
play. The test waits quite a bit longer (4 "dmc_cycles"), and makes sure
that the dmc_en flag is clear.

The tricky part (for me), was realizing that the sample buffer needed 3
states. If you just implement it with 2 states (has/doesn't have byte)
then it doesn't work properly. The "has sample" flag should be set as
long as there is a sample byte playing. But if that's true then we can't
tell whether we should clear the "has sample" flag when copying the
sample byte into the shift register. The way to know is to track whether
the sample byte has been played yet.