Implement undocumented instructions
ANC: AND + ASL/ROL
ARR: AND + LSR
ANE: TXA + AND #nn (but with weird quirks)
LXA: load A X (but with weird quirks)
SHA: store A & X & H
SHS: store A & X in S and memory
SHY: store Y & H
SHX: store X & H
AXS: X = A & X - #imm

Not sure if their implementations are completely correct, but (most of)
the tests pass. Most of the implementations are using the "Rarely used
operations" bit.