Skip DMC fetch steps if one is already happening
If DMC is enabled right before a DMC sample byte finishes, then the a
fetch will be queued during the fetch steps. This will lock up the
emulator, since the "dmc_step" (the step to return to after DMC fetch is
finished) will be overwritten. This happened reliably in the TMNT intro.