Fix sprite_ram test
When performing OAM DMA, 256 bytes are copied from memory. The high
bytes of the source address is the value written to $4014, and the low
byte always starts at 0. The destination address is always OAM, but the
offset begins with the value written to $2003. This was incorrectly
implemented before, but now is fixed.