Implement mappers 21, 22, 23, and 25
These are the VRC2 and VRC4 variants. VRC4 has a built-in IRQ, so I've
renamed `IRQ_MMC3` to `IRQ_MAPPER`.

The maximum bank index is now > 256 too so it can't be `u8`; I've
updated all of them to be `u16` now.

Determining which mapper behavior to use from a combination of the NES
header, NES 2.0 header, and the cart database is pretty clunky and could
use a cleanup.