Fix VRC mappers not in cartdb
Some of the various VRC2/VRC4 configurations are assigned to the same
mappers, so we need to try to handle mapper register writes for multiple
addresses.

Also, fix the fourscreen flag so it is only enabled in some mappers;
ignore_mirror doesn't always imply four screen mode.

These changes fix the Axelay demo.