Optimize `shift_en()` some more
Remove `u32x4 shifter` and replace it with `u32 bgatshift`, which stores
the full 4-bit palette index for each of the 8 pixels. This means that
`shift_en()` doesn't have to mix the bits, but instead can just mask of
the lowest 4.

Also bake the left mask into the BG and sprite shifters, so we can avoid
an additional mask in `shift_en()`.

Interestingly, this ends up not being much faster for native builds (or
perhaps even slightly slower), but seems to save ~1ms per frame on Wasm
builds.