Fix test28, implement reset playback
This change was surprisingly involved. We don't want the emulator to
know about `Joypad`, and we want the reset button to be encoded in the
joypad playback. The joypad playback is only read/advanced during joypad
reading, which means that a ROM that never reads the joypad can never be
reset.

The solution is somewhat elegant and clumsy at the same time; we use the
emulator's existing scheduler to schedule when the reset will occur
during playback. This will allow the emulator to cleanly exit with an
`EmulatorEvent` when the reset occurs. When this happens the host will
trigger the reset, and then schedule the next reset event, if any.

This change also fixes `test28.nes`, and the
`cpu_reset/ram_after_reset.nes` and `cpu_reset/registers.nes` tests.
This requires properly implementing reset, which should behave like IRQ
or NMI, but not perform any writes to the stack.