Add reset to movie files
The movie file is a collection of frames. Each frame contains a latch value
and an input value. The latch is a count of how many times the input must
be read by the game before the input is returned.

If the latch contains its high bit set, then this instead represents a reset
marker, and the latch value is the number of CPU cycles since the last reset.
It's possible that this doesn't correspond with the current input, which
would indicate that the file probably won't work.