Fix mmc3 irq prediction crash w/ Mickey games
The bug in both Mickey in Numberland/Letterland occurs because normally an IRQ
clock wouldn't occur at that cycle (i.e. the low count is not long enough). But
that is assuming that things are executing normally. If rendering was just
enabled, then it's possible to have a much longer "low" count, so the next A12
high will clock the IRQ counter.